Move pipeline init to separate file
